Refactor image upload component out of ChatDialogInputRow
